Removing Data Files:
Since we already had a lot of the data files in the repo, I had to delete them manually.
Just modifying the .gitinore would not ignore the files since they were already uploaded.
What I did is, I modifed the .gitignore so that future data files will not be added to the repo.
Since I had to removed the files manually, I recommend that you save them locally before pulling the changes.
Then you can add them and they will not be included in the next push.